extract timeout
