+++ test/decorateUserResHeaders.js
@@ -4,10 +4,11 @@ var assert = require('assert');
 var express = require('express');
 var request = require('supertest');
 var proxy = require('../');
+var proxyTarget = require('./support/proxyTarget');
+var TIMEOUT = require('./constants');
 
 describe('when userResHeaderDecorator is defined', function () {
-
-  this.timeout(10000);
+  this.timeout(TIMEOUT.STANDARD);
 
   var app;
   var  serverReference;
