+++ test/https.js
@@ -5,9 +5,10 @@ var express = require('express');
 var request = require('supertest');
 var proxy = require('../');
 var proxyTarget = require('./support/proxyTarget');
+var TIMEOUT = require('./constants');
 
 describe('proxies https', function () {
-  this.timeout(10000);
+  this.timeout(TIMEOUT.STANDARD);
 
   var app;
   var proxyServer;
