+++ test/preserveHostHdr.js
@@ -4,8 +4,9 @@ var assert = require('assert');
 var express = require('express');
 var request = require('supertest');
 var proxy = require('../');
+var proxyTarget = require('./support/proxyTarget');
+var TIMEOUT = require('./constants');
 
-var proxyTarget = require('../test/support/proxyTarget');
 var proxyRouteFn = [{
   method: 'get',
   path: '/hostHdrTest',
@@ -16,7 +17,7 @@ var proxyRouteFn = [{
 
 describe('preserves host header only when requested', function () {
 
-  this.timeout(10000);
+  this.timeout(TIMEOUT.STANDARD);
 
   var app;
   var proxyServer;
